requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database dependencies
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0

# AI and HTTP dependencies
google-generativeai==0.3.2
httpx==0.25.2
aiohttp==3.9.1

# Message Queue dependencies
redis==5.0.1
celery==5.3.4
aio-pika==9.3.1
pydantic==2.5.2

# Utilities
python-dateutil==2.8.2
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1